github(service): combine branches and PR mixins into a single module\n\n- New module: service/branches_prs.py contains GitHubBranchesMixin and GitHubPRsMixin\n- Update service/__init__.py to import both from combined module\n- Remove old service/branches.py and service/prs.py\n\nCo-authored-by: openhands <openhands@all-hands.dev>
