+++ openhands/integrations/github/service/__init__.py
@@ -1,6 +1,9 @@
 from pydantic import SecretStr
 
-from openhands.integrations.github.service.branches import GitHubBranchesMixin
+from openhands.integrations.github.service.branches_prs import (
+    GitHubBranchesMixin,
+    GitHubPRsMixin,
+)
 from openhands.integrations.github.service.core import GitHubCoreMixin
 from openhands.integrations.github.service.graphql import GitHubGraphQLMixin
 from openhands.integrations.github.service.http import GitHubHTTPMixin
@@ -8,7 +11,6 @@ from openhands.integrations.github.service.installations import (
     GitHubInstallationsMixin,
 )
 from openhands.integrations.github.service.microagents import GitHubMicroagentsMixin
-from openhands.integrations.github.service.prs import GitHubPRsMixin
 from openhands.integrations.github.service.repos import GitHubReposMixin
 from openhands.integrations.service_types import (
     BaseGitService,
